feat: Day11p2 solved. Issue was overflow again!!!
The number I had been getting over and over was a 2s
complement of the correct answer! The reason rust
wasn't complaining is because I kept running --develop.

Ugh I'm so mad. The graph did NOT have cycles and the
naive approach I thought of a while ago of multiplying
the number of subpaths was fine